third_party/blink/web_tests/platform/generic/media/W3C/audio/preload/preload_reflects_auto_value-expected.txt

audio.preload - reflection test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "auto" is "auto"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
